todo_drf/requirements.txt
asgiref==3.3.1
certifi==2020.12.5
cffi==1.14.5
chardet==4.0.0
cryptography==3.4.6
defusedxml==0.7.0
Django==3.1.7
django-allauth==0.44.0
django-rest-auth==0.9.5
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.6.0
drf-base64==2.0
idna==2.10
jsonfield==3.1.0
oauthlib==3.1.0
Pillow==8.1.2
psycopg2==2.8.6
pycparser==2.20
PyJWT==1.7.1
python3-openid==3.2.0
pytz==2019.3
requests==2.25.1
requests-oauthlib==1.3.0
six==1.15.0
sqlparse==0.3.0
urllib3==1.26.3
wincertstore==0.2
